Refactor cloud_backup_manager: move workers to actors
Reorganize the cloud backup manager actor code by introducing an actors module and moving/renaming the previous workers files into it. New files and moves: created rust/src/manager/cloud_backup_manager/actors.rs and actors submodules (cleanup, restore, supervisor, sync_health, uploads, write), split write into supervisor and worker, and relocated many workers/* -> actors/*. Updated imports and re-exports across the crate to use self::actors::{...} (and adjusted several super:: paths), added a CloudBackupWriteError alias, and updated tests and ops to reference the new module layout. This change separates actor components for clearer module structure and better encapsulation of supervisor/worker responsibilities.
