+++ rust/src/manager/cloud_backup_manager/actors.rs
@@ -1,0 +1,15 @@
+pub(crate) mod cleanup;
+pub(crate) mod restore;
+pub(crate) mod supervisor;
+mod sync_health;
+mod uploads;
+pub(crate) mod write;
+
+pub(crate) use self::cleanup::{CleanupExpectedWalletRecord, CleanupSourceNamespace};
+pub(crate) use self::restore::CloudBackupRestoreEvent;
+pub(crate) use self::supervisor::{CloudBackupOperation, CloudBackupSupervisor};
+pub(crate) use self::write::{
+    CloudBackupUploadedWallet, CloudBackupWalletCountRefresh, CloudBackupWriteBlocker,
+    CloudBackupWriteClient, CloudBackupWriteCompletion, CloudBackupWriteResultReceiver,
+    CloudBackupWriteSupervisor,
+};
