+++ rust/src/manager/cloud_backup_manager/actors/write.rs
@@ -1,0 +1,10 @@
+mod supervisor;
+mod worker;
+
+pub(crate) use self::supervisor::{
+    CloudBackupUploadedWallet, CloudBackupUploadedWalletsStateMode, CloudBackupWalletCountRefresh,
+    CloudBackupWriteBlocker, CloudBackupWriteClient, CloudBackupWriteCommandResult,
+    CloudBackupWriteCompletion, CloudBackupWriteResultReceiver, CloudBackupWriteSupervisor,
+};
+
+pub(crate) type CloudBackupWriteError = crate::manager::cloud_backup_manager::CloudBackupError;
